+++ lib/plugins/index.js
@@ -14,7 +14,7 @@ module.exports = [
   require('./print.js'),
   require('./remove.js'),
   require('./rollback.js'),
-  require('./slstats/slstats.js'),
+  require('./slstats.js'),
   require('./plugin/plugin.js'),
   require('./plugin/install.js'),
   require('./plugin/uninstall.js'),
