+++ lib/plugins/slstats.js
